move to monorepo structure
