+++ package.json
@@ -1,30 +1,9 @@
 {
-	"name": "toolbox-devvit",
-	"version": "0.3.2",
-	"description": "Helpers for working with /r/toolbox data from Devvit community apps.",
-	"type": "module",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
-	"files": [
-		"dist/*"
+	"private": true,
+	"workspaces": [
+		"packages/toolbox-devvit"
 	],
-	"scripts": {
-		"fmt": "dprint fmt",
-		"build": "tsc",
-		"test": "ava",
-		"coverage": "nyc ava",
-		"docs": "typedoc src/index.ts",
-		"prepublishOnly": "tsc && ava"
-	},
-	"repository": "https://github.com/toolbox-team/toolbox-devvit.git",
-	"author": "eritbh <github@eritbh.me>",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/toolbox-team/toolbox-devvit/issues"
-	},
-	"homepage": "https://github.com/toolbox-team/toolbox-devvit#readme",
 	"devDependencies": {
-		"@ava/typescript": "^4.0.0",
 		"@devvit/public-api": "^0.11.2",
 		"@tsconfig/node-lts": "^22.0.1",
 		"@types/node": "^13.7.4",
@@ -32,20 +11,7 @@
 		"dprint": "^0.40.2",
 		"eslint": "^6.8.0",
 		"nyc": "^15.0.0",
-		"tsx": "^3.12.7",
 		"typedoc": "^0.24.8",
 		"typescript": "^5.1.3"
-	},
-	"dependencies": {
-		"pako": "^1.0.11"
-	},
-	"ava": {
-		"extensions": {
-			"ts": "module"
-		},
-		"nodeArguments": [
-			"--loader=tsx",
-			"--no-warnings"
-		]
 	}
 }
