+++ packages/toolbox-devvit/.npmignore
@@ -1,0 +1,4 @@
+src
+tsconfig.json
+dist/**/*.test.{js,d.ts}
+*.tsbuildinfo
