+++ packages/toolbox-devvit/package.json
@@ -1,0 +1,34 @@
+{
+	"name": "toolbox-devvit",
+	"version": "0.3.2",
+	"description": "Helpers for working with /r/toolbox data from Devvit community apps.",
+	"type": "module",
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"scripts": {
+		"fmt": "dprint fmt",
+		"build": "tsc",
+		"test": "ava",
+		"coverage": "nyc ava",
+		"docs": "typedoc src/index.ts",
+		"prepublishOnly": "tsc && ava"
+	},
+	"repository": "https://github.com/toolbox-team/toolbox-devvit.git",
+	"author": "eritbh <github@eritbh.me>",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/toolbox-team/toolbox-devvit/issues"
+	},
+	"homepage": "https://github.com/toolbox-team/toolbox-devvit#readme",
+	"dependencies": {
+		"pako": "^1.0.11"
+	},
+	"ava": {
+		"typescript": {
+			"compile": "tsc",
+			"rewritePaths": {
+				"src/": "dist/"
+			}
+		}
+	}
+}
