+++ packages/toolbox-devvit/tsconfig.json
@@ -13,8 +13,5 @@
 	},
 	"include": [
 		"src"
-	],
-	"exclude": [
-		"src/**/*.test.ts"
 	]
 }
